Added last remaining instructions to the parser

+++ modules/core/include/miniqbt/emulator/ControlledNotGate.hpp
@@ -24,8 +24,8 @@ namespace MiniQbt{
         *   @param state,               the state that will be altered.
         *   @throws QuantumException,   when the an invalid bit is altered it will throw a quantum exception
         */
-        template<size_t registerSize>
-        void cnotGate(const size_t& control,const size_t& target, QuantumState<registerSize>& state){
+        template<int registerSize>
+        void cnotGate(const int& control,const int& target, QuantumState<registerSize>& state){
             using namespace Tools;
             assertInput(control == target, "The control bit can't be the same as the target bit...");
             assertInput(target >= registerSize, "Target bit can't be bigger then the register.");
@@ -35,8 +35,7 @@ namespace MiniQbt{
             for(size_t i = 0; i < order[1].size(); i++){
                 order[1][i].bitflip(target); 
             }
-            
-            
+
             state = mergeOrderProb<registerSize>(order[0], order[1]);
         }
     }
